Removed out-of-bounds anchor in Tile 301A causing tile displacement
The group center was getting averaged with an anchor that was at X position 115, causing any Tile Group with tile 301A in it to be way off the normal map, making it look like the original map had disappeared. I tried searching for similar displacements in other tiles via regex on their prefabs and didn't find anything.