Enable skin-switching and multiple random skin variants per monster
When switching skins, it sets the overrideSprite on the MonsterButton and on the CombatPanel. Also updated the Settings Dialog to have the Skinpack dropdown selector. And updated the [x] close button to an [OK] button at the bottom. Also tinted the [Quit to Title] button red to make it less easy to push and quit the scenario.